Move dependency tests to lib
